Support IoTEdgeHubDev simulator as host
EdgeHubConnectionString environment variable is set by IoTEdgeHubDev tooling
to pass CS to a module.  Without it publisher fails to start under simulation.

Fix: Check for environment variable to be present and thus declare it to be
running int edge context.

The ModuleClient.CreateFromEnvironmentVariable will first check the
EdgeHubConnectionString environment variable before using the individual
variables thus the simple check for it to be defined will cause the module
client to be used.  Thus no more work needed.